Don't fail first stage mount if fstab is not found
The first stage mount looks for default fstab, but we rely on properties
to find fstab.${ro.hardware}. However properties are not available yet
in that stage. Prior to Android S, failure to find default fstab at that
stage was non fatal. As we don't need the first stage mount, just make
it non fatal again.

Change-Id: Ibaf1d9d1c278e2dcacbbd149f3661e00c3cc1553

Former-commit-id: 23c65b99441a1f431d592f8a7d2927f69071c46a
Former-commit-id: 197a066858332e64cc18a422e87e52bf5fed8bf5