healthd: Some improvement to BatteryMonitor
- Allow batteryFullCharge, batteryFullChargeDesignCapacity and
batteryChargeCounter to look for energy_* info under the battery's
sysfs if it's available
Different from charge_*, these energy are represent in µWh instead
of µAh according to the Linux Kernel Documentation :
https://docs.kernel.org/power/power_supply_class.html

So we will have to convert it into µAh, take voltage_min_design as
the base voltage and calculate.

- Set a property to get batteryFullChargeDesignCapacityUah value,
we will modify PowerProfile to get the capacity with this property.

Change-Id: I806ac5d0b8caed2fed86e01c20d1f2da7f4acf96
Signed-off-by: hmtheboy154 <buingoc67@gmail.com>